implement calculateMusicalProfile method
implemented calculateMusicalProfile method which is used in both user login and the algorithm
