+++ backend/Controllers/loginController.js
@@ -3,6 +3,7 @@ require("dotenv").config();
 
 let User = require("../Models/user.model");
 let Group = require("../Models/group.model");
+const { calculateMusicalProfile } = require("../scripts");
 
 const client_id = process.env.CLIENT_ID; // Your client id
 const client_secret = process.env.CLIENT_SECRET; // Your secret
@@ -44,7 +45,7 @@ exports.loginUser = async (req, res) => {
       //get user's top 50 tracks
       try {
         //spotify api call
-        let data = await spotifyApi.getMyTopTracks({ limit: 50, time_range:"med_term" });
+        let data = await spotifyApi.getMyTopTracks({ limit: 50 });
 
         for (x of data.body.items) {
           let track = {};
@@ -108,7 +109,7 @@ exports.loginUser = async (req, res) => {
           i++; // iterate for the next item to add in our topTracks array
         }
       } catch (err) {
-        res.json({ message: "Unable to get track audio featurs.", error: err });
+        res.json({ message: "Unable to get track audio features.", error: err });
         return;
       }
 
@@ -141,31 +142,9 @@ exports.loginUser = async (req, res) => {
         return;
       }
 
-      // initiate vars for the musical profile
-      let pop = (dnce = nrgy = spch = acst = inst = vale = 0);
-
-      // Calculate their musical profile (use averages for now) (calc sums and div n)
-      for (x of topTracks) {
-        pop += x.trackPopularity;
-        dnce += x.danceability;
-        nrgy += x.energy;
-        spch += x.speechiness;
-        acst += x.acousticness;
-        inst += x.instrumentalness;
-        vale += x.valence;
-      }
-
-      topTracksLength = topTracks.length;
-
-      musicalProfile = {
-        popularity: pop / topTracksLength,
-        danceability: (dnce / topTracksLength) * 100,
-        energy: (nrgy / topTracksLength) * 100,
-        speechiness: (spch / topTracksLength) * 100,
-        acousticness: (acst / topTracksLength) * 100,
-        instrumentalness: (inst / topTracksLength) * 100,
-        valence: (vale / topTracksLength) * 100,
-      };
+      // use scripted method for calculating musical profile and store in musicalProfile
+      
+      // musicalProfile = calculateMusicalProfile(topTracks);
 
       //get user object from SpotifyAPI
       spotifyApi.getMe().then(
@@ -180,7 +159,7 @@ exports.loginUser = async (req, res) => {
                 name: data.body.display_name,
                 imageURL: data.body.images[0].url,
                 email: data.body.email,
-                musicalProfile: musicalProfile,
+                musicalProfile: calculateMusicalProfile(topTracks),
                 topTracks: topTracks, // Add top 50 tracks with their attributes
                 topArtists: topArtists, // Add top 30 artists with their attributes
               },
