+++ backend/scripts.js
@@ -18,5 +18,42 @@ function generateGroupCode(){
     return password
 }
 
-//export generate() as a module for use in Controller.js
+/**
+ * Method used to calculate the averages of people's top artists
+ * @param {[track objects]} data, pass array of track objects to calcualte the averages 
+ * @returns 
+ */
+function calculateMusicalProfile (data) {
+    let pop = (dnce = nrgy = spch = acst = inst = vale = 0);
+
+    for (x of data){
+        pop += x.trackPopularity;
+        dnce += x.danceability;
+        nrgy += x.energy;
+        spch += x.speechiness;
+        acst += x.acousticness;
+        inst += x.instrumentalness;
+        vale += x.valence;
+    }
+
+    dataLen = data.length;
+
+    musicalProfile = {
+        popularity: pop / dataLen,
+        danceability: (dnce / dataLen) * 100,
+        energy: (nrgy / dataLen) * 100,
+        speechiness: (spch / dataLen) * 100,
+        acousticness: (acst / dataLen) * 100,
+        instrumentalness: (inst / dataLen) * 100,
+        valence: (vale / dataLen) * 100,
+      };
+
+    return musicalProfile;
+}
+
+
+//export music profile calculation method
+module.exports.calculateMusicalProfile = calculateMusicalProfile;
+
+//export generateGroupCode() as a module for use in Controller.js
 module.exports.generateGroupCode = generateGroupCode;
